+++ packages/expo-modules-core/CHANGELOG.md
@@ -16,6 +16,7 @@
 ### 💡 Others
 
 - [iOS] Improved conversions of returned arrays and dictionaries with mixed element types. ([#42641](https://github.com/expo/expo/pull/42641) by [@barthap](https://github.com/barthap))
+- [iOS] Make RNHostView SwiftUI view ([#43570](https://github.com/expo/expo/pull/43570) by [@nishan](https://github.com/intergalacticspacehighway))
 
 ## 55.0.12 — 2026-02-25