+++ packages/expo-ui/CHANGELOG.md
@@ -17,6 +17,7 @@
 ### 💡 Others
 
 - [iOS] Introduce `SlotView` to replace structural child view types with a single generic slot. ([#43607](https://github.com/expo/expo/pull/43607) by [@nishan](https://github.com/intergalacticspacehighway))
+- [iOS] Make RNHostView SwiftUI view ([#43570](https://github.com/expo/expo/pull/43570) by [@nishan](https://github.com/intergalacticspacehighway))
 
 ## 55.0.1 — 2026-02-25